// src/stories/components/StoryComponent/enums.ts
export enum ContentAlignItems {
  top = 'top',
  center = 'center',
  bottom = 'bottom',
}

export enum ContentDisplay {
  flex = 'flex',
  grid = 'grid',
}

export enum ContentGridFlow {
  column = 'column',
  row = 'row',
  maxTwoColumns = 'max-two-columns',
  maxThreeColumns = 'max-three-columns',
  maxFourColumns = 'max-four-columns',
}