// src/shared/ComponentsUI/components/Tooltip/enums.ts
export enum TooltipPosition {
  topStart = 'top-start',
  topCenter = 'top-center',
  topEnd = 'top-end',
  bottomStart = 'bottom-start',
  bottomCenter = 'bottom-center',
  bottomEnd = 'bottom-end',
  leftStart = 'left-start',
  leftCenter = 'left-center',
  leftEnd = 'left-end',
  rightStart = 'right-start',
  rightCenter = 'right-center',
  rightEnd = 'right-end',
}